requirements.txt
haystack-ai>=2.0.0
sentence-transformers>=5.0.0
openai>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
streamlit>=1.32.0